made a call to log someone out (invalidates their server side session) made it visible if the user's account is disabled (needs work but it is a start) fixed issues with zooming in the image map/browser changed with width of the lists  to not be 100% added a role class for future user roles deleted the old bulk upload classes, and renamed the new ones disabled adding images to the grid for bulk uploaded images fixed an issue where you could not do forgot password

+++ mpdb-client/src/edu/rpi/metpetdb/client/ui/image/browser/ImageBrowserMouseListener.java
@@ -395,7 +395,7 @@ public class ImageBrowserMouseListener implements MouseListener {
 		// we multiple by -1, because if we are zoomed out the images
 		// are
 		// smaller, but we have to multiply by the scale when moving
-		final int scale = imageBrowser.getZoomScale();
+		final float scale = imageBrowser.getZoomScale();
 		GWT.log("Scale:" + scale + " deltaX:" + deltaX + " deltaY:" + deltaY
 				+ " currentX:" + currentImage.getCurrentContainerPosition().x
 				+ " currentY:" + currentImage.getCurrentContainerPosition().y
@@ -508,7 +508,7 @@ public class ImageBrowserMouseListener implements MouseListener {
 			deltaX = x - lastX;
 		}
 
-		final int scale = imageBrowser.getZoomScale() * -1 == 0 ? 1
+		final float scale = imageBrowser.getZoomScale() * -1 == 0 ? 1
 				: imageBrowser.getZoomScale() * -1;
 		currentImage.move(deltaX, deltaY, scale);
 		currentImage.getIog()
